fix(examples): some platform not support double

+++ examples/vadd/main.cpp
@@ -1,7 +1,7 @@
 #include <sycl/sycl.hpp>
 
 extern "C" SYCL_EXTERNAL
-void vector_add(double* a, double* b) {
+void vector_add(float* a, float* b) {
     auto item = sycl::ext::oneapi::this_work_item::get_nd_item<1>();
     int idx = item.get_global_id(0);
